--- src/main/java/com/heroku/backend/controller/HealthCheckController.java
@@ -1,20 +1,0 @@
-package com.heroku.backend.controller;
-
-import com.heroku.backend.service.HealthCheckService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HealthCheckController {
-    public final HealthCheckService healthCheckService;
-
-    public HealthCheckController(HealthCheckService healthCheckService) {
-        this.healthCheckService = healthCheckService;
-    }
-
-    @GetMapping("/healthcheck")
-    public ResponseEntity<String> healthCheck() {
-        return healthCheckService.healthCheck();
-    }
-}
