--- src/main/java/com/heroku/backend/exception/InvalidBase64StringException.java
@@ -1,11 +1,0 @@
-package com.heroku.backend.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(code = HttpStatus.FORBIDDEN, reason = "Request rejected due to invalid bae64 login string")
-public class InvalidBase64StringException extends Exception {
-    public InvalidBase64StringException() {
-        super();
-    }
-}
