--- src/main/java/com/heroku/backend/repository/TaskRepository.java
@@ -1,8 +1,0 @@
-package com.heroku.backend.repository;
-
-import com.heroku.backend.entity.TaskEntity;
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface TaskRepository extends MongoRepository<TaskEntity, String> {
-
-}
