+++ src/main/java/de/janetschel/haushaltsplan/backend/App.java
@@ -1,0 +1,11 @@
+package de.janetschel.haushaltsplan.backend;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class App {
+    public static void main(String[] args) {
+        SpringApplication.run(App.class, args);
+    }
+}
