+++ src/main/java/de/janetschel/haushaltsplan/backend/repository/TaskRepository.java
@@ -1,0 +1,8 @@
+package de.janetschel.haushaltsplan.backend.repository;
+
+import de.janetschel.haushaltsplan.backend.entity.TaskEntity;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface TaskRepository extends MongoRepository<TaskEntity, String> {
+
+}
